+++ shardingsphere-features/shardingsphere-readwrite-splitting/shardingsphere-readwrite-splitting-distsql/shardingsphere-readwrite-splitting-distsql-handler/src/main/java/org/apache/shardingsphere/readwritesplitting/distsql/handler/exception/InvalidLoadBalancersException.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.readwritesplitting.distsql.exception;
+package org.apache.shardingsphere.readwritesplitting.distsql.handler.exception;
 
 import org.apache.shardingsphere.infra.exception.rule.RuleDefinitionViolationException;
