+++ shardingsphere-features/shardingsphere-readwrite-splitting/shardingsphere-readwrite-splitting-distsql/shardingsphere-readwrite-splitting-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.query.RQLResultSet
@@ -15,6 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.readwritesplitting.distsql.update.CreateReadwriteSplittingRuleStatementUpdater
-org.apache.shardingsphere.readwritesplitting.distsql.update.AlterReadwriteSplittingRuleStatementUpdater
-org.apache.shardingsphere.readwritesplitting.distsql.update.DropReadwriteSplittingRuleStatementUpdater
+org.apache.shardingsphere.readwritesplitting.distsql.handler.query.ReadwriteSplittingRuleQueryResultSet
